Add files via upload
Example use of V1 of the rankings and team selector code. To use yourself, download all of the files in "Rankings_folder_1" and set the relevant paths to the rankings folder for your computer in "WEIGHTLIFTING RANKINGS BUILDER AND TEAM SELECTOR (V1) Example.ipynb". You should get the same results as in the excited cells of the code as long as your paths are correct and you have all of the libraries downloaded at the top of the script.