Add files via upload
Executed version of the WRBTS.ipynb script with all needed input and output sheets. To reproduce results on your computer, download all items in the "WRBTS V1 Executed" folder and set the correct paths to the "Rankings_Folder_1" file in the script, and make sure you have the correct libraries downloaded. Then run the script and it should work.